commit modify

+++ NewCRM.Domain/Entitys/System/App.cs
@@ -126,7 +126,7 @@ namespace NewCRM.Domain.Entitys.System
 
         public Boolean IsInstall { get; private set; }
 
-        public Int32 StarCount { get; private set; }
+        public Double StarCount { get; private set; }
 
         public String AccountName { get; set; }
